feat: write a hook method for import logpath

+++ tests/conftest.py
@@ -1,16 +1,27 @@
 import pytest,os
 from core import parse
 
+def pytest_addoption(parser):
+    parser.addoption(
+        "--logpath",
+        action="store",
+        default=None,
+        help="input a logpath to analyze"
+    )
+
+
 @pytest.fixture(scope="session")
-def logexport():
-    # 通过adb 命令我估计会用subprocess.run（[adb bugreport]）来导出设备日志，这里用打开创建文件方式mock
-    with open("app.log",'w',encoding='utf-8') as f:
-        f.write('FATAL EXCEPTION:1\n')
-        f.write('memory leak:2\n')
-    dummy_logs = "./app.log"
-    yield dummy_logs
-    # 这里我不确定日志是否清理
-    os.remove(dummy_logs)
+def logexport(request):
+    log_path = request.config.getoption("--logpath")
+    if log_path is None:
+        with open("app.log",'w',encoding='utf-8') as f:
+            f.write('FATAL EXCEPTION:1\n')
+            f.write('memory leak:2\n')
+        dummy_logs = "./app.log"
+        yield dummy_logs
+        os.remove(dummy_logs)
+    else:
+        yield log_path
 
 @pytest.fixture(scope="session")
 def logparse(logexport):
